refactor: move a helper file into builder
